+++ lib/src/constants/local_storage_constants.dart
@@ -2,7 +2,5 @@
 
 /// Local database keys.
 class DBKeys {
-  static const String P1 = "P01_";
-  static const String P2 = "P02_";
-  static const String P3 = "P03_";
+  static const String CIPHER_ALGORITHM_INDEX = "CIPHER_ALGORITHM_INDEX";
 }