+++ lib/src/core/_core.dart
@@ -1,5 +1,5 @@
-export 'app/app_bootstrap.dart';
-export 'app/app_bootstrap_container.dart';
+export 'app/app_startup.dart';
+export 'app/app_startup_container.dart';
 export 'app/navigation.dart';
 export 'app/shared_prefs.dart';
 export 'network/api_service.dart';