+++ lib/src/features/cipher/data/local/local_cipher_repo.dart
@@ -1,4 +1,6 @@
 import 'package:blowfish/blowfish.dart';
+import 'package:cipher_dove/src/constants/_constants.dart';
+import 'package:cipher_dove/src/core/_core.dart';
 import 'package:cipher_dove/src/features/cipher/domain/cipher_action.dart';
 import 'package:cipher_dove/src/features/cipher/domain/cipher_algorithm.dart';
 import 'package:cipher_dove/src/features/cipher/domain/cipher_mode.dart';
@@ -6,23 +8,34 @@ import 'package:cryptography_plus/cryptography_plus.dart';
 import 'package:crypto/crypto.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-// import 'package:sha3/sha3.dart';
+import 'package:sha3/sha3.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 part 'local_cipher_repo.g.dart';
 
 // TODO abstraction
 class LocalCipherRepository {
   const LocalCipherRepository(
+    this.sharedPref,
     this.crypt,
-    // this.sha3Factory,
+    this.sha3Factory,
   );
 
+  final SharedPreferencesAsync sharedPref;
   final Cryptography crypt;
-  // final SHA3 sha3Factory;
+  final SHA3 sha3Factory;
 
   static const keyLength = 32;
 
-  // TODO warning user about key must be 32 char.
+  Future<CipherAlgorithm> getDefaultAlgorithm() async {
+    final index = await sharedPref.getInt(DBKeys.CIPHER_ALGORITHM_INDEX);
+    return CipherAlgorithm.values[index ?? Default.CIPHER_ALGORITHM_INDEX];
+  }
+
+  Future<void> setDefaultAlgorithm(CipherAlgorithm algorithm) async {
+    return await sharedPref.setInt(DBKeys.CIPHER_ALGORITHM_INDEX, CipherAlgorithm.values.indexOf(algorithm));
+  }
+
   String padKey(String key) {
     if (key.length < keyLength) return key.padRight(keyLength, '0'); // Padding
     if (key.length > keyLength) return key.substring(0, keyLength); // Truncating
@@ -72,7 +85,7 @@ class LocalCipherRepository {
   /// Will always decrypt if not encrypt.
   Future<String> chacha20(String input, SecretKey secretKey, {required CipherAction action}) async {
     final chacha20 = crypt.chacha20(macAlgorithm: MacAlgorithm.empty);
-    final nonce = (input.length < 12) ? input.padRight(12, '0') : input;
+    final nonce = '0'.padRight(12, '0');
 
     if (action == CipherAction.encrypt) {
       final encrypted = await chacha20.encrypt(
@@ -97,7 +110,7 @@ class LocalCipherRepository {
   /// Will always decrypt if not encrypt.
   Future<String> aes(String input, SecretKey secretKey, {required CipherAction action}) async {
     final aes = crypt.aesCbc(macAlgorithm: MacAlgorithm.empty);
-    final nonce = (input.length < 16) ? input.padRight(16, '0') : input;
+    final nonce = '0'.padRight(16, '0');
 
     if (action == CipherAction.encrypt) {
       final encrypted = await aes.encrypt(
@@ -122,7 +135,7 @@ class LocalCipherRepository {
   /// Will always decrypt if not encrypt.
   Future<String> blowfish(String input, String secretKey, {required CipherAction action}) async {
     final blowfish = Blowfish();
-    final iv = (input.length < 8) ? input.padRight(8, '0') : input;
+    final iv = '0'.padRight(8, '0');
 
     if (action == CipherAction.encrypt) {
       final encrypted = blowfish.encryptCBC(input.codeUnits, iv.codeUnits.sublist(0, 8), applyPadding: true);
@@ -139,7 +152,11 @@ class LocalCipherRepository {
 //
 
   // TODO asymmetric support
-  String asymmetric(String input, {required CipherMode mode}) {
+  Future<String> encryptAsymmetric(String input, {required CipherMode mode}) async {
+    return "";
+  }
+
+  Future<String> decryptAsymmetric(String input, {required CipherMode mode}) async {
     return "";
   }
 
@@ -152,7 +169,7 @@ class LocalCipherRepository {
       == CipherAlgorithm.md5 => md5.convert(input.codeUnits).toString(),
       == CipherAlgorithm.sha1 => await sha1(input),
       == CipherAlgorithm.sha2 => await sha2(input),
-      // == CipherAlgorithm.sha3 => await sha3(input),
+      == CipherAlgorithm.sha3 => await sha3(input),
       == CipherAlgorithm.blake2 => await blake2(input),
       _ => throw UnsupportedError("CipherAlgorithm not supported."),
     };
@@ -176,13 +193,13 @@ class LocalCipherRepository {
     return digest.bytes.map((b) => b.toRadixString(16).padLeft(2, '0')).join();
   }
 
-  // Future<String> sha3(String input) async {
-  //   sha3Factory.update(input.codeUnits);
+  Future<String> sha3(String input) async {
+    sha3Factory.update(input.codeUnits);
 
-  //   final digest = sha3Factory.digest();
-  //   // Convert the digest to a hexadecimal string
-  //   return digest.map((byte) => byte.toRadixString(16).padLeft(2, '0')).join();
-  // }
+    final digest = sha3Factory.digest();
+    // Convert the digest to a hexadecimal string
+    return digest.map((byte) => byte.toRadixString(16).padLeft(2, '0')).join();
+  }
 
   Future<String> blake2(String input) async {
     final algorithm = crypt.blake2b();
@@ -198,10 +215,9 @@ class LocalCipherRepository {
 
 @riverpod
 LocalCipherRepository localCipherRepository(Ref ref) {
+  final sharedPref = ref.watch(sharedPrefProvider);
   final crypt = Cryptography.defaultInstance;
-  // final sha3Factory = SHA3(256, KECCAK_PADDING, 256);
-  return LocalCipherRepository(
-    crypt,
-    // sha3Factory,
-  );
+  final sha3Factory = SHA3(256, KECCAK_PADDING, 256);
+
+  return LocalCipherRepository(sharedPref, crypt, sha3Factory);
 }