+++ lib/src/features/cipher/presentation/cipher_output_controller.g.dart
@@ -1,26 +1,27 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'cipher_mode_state.dart';
+part of 'cipher_output_controller.dart';
 
 // **************************************************************************
 // RiverpodGenerator
 // **************************************************************************
 
-String _$cipherModeStateHash() => r'aa76dd30f413e5e9ef9de5410c3cee0f3bed604d';
+String _$cipherOutputControllerHash() =>
+    r'2000c69c375d5f7ab30195a13f05ee2a6ed5ff8c';
 
-/// See also [CipherModeState].
-@ProviderFor(CipherModeState)
-final cipherModeStateProvider =
-    NotifierProvider<CipherModeState, CipherMode>.internal(
-  CipherModeState.new,
-  name: r'cipherModeStateProvider',
+/// See also [CipherOutputController].
+@ProviderFor(CipherOutputController)
+final cipherOutputControllerProvider =
+    AsyncNotifierProvider<CipherOutputController, String>.internal(
+  CipherOutputController.new,
+  name: r'cipherOutputControllerProvider',
   debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
       ? null
-      : _$cipherModeStateHash,
+      : _$cipherOutputControllerHash,
   dependencies: null,
   allTransitiveDependencies: null,
 );
 
-typedef _$CipherModeState = Notifier<CipherMode>;
+typedef _$CipherOutputController = AsyncNotifier<String>;
 // ignore_for_file: type=lint
 // ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member, deprecated_member_use_from_same_package